fix after code review
