+++ src/App.tsx
@@ -1,18 +1,25 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { UserWarning } from './UserWarning';
 import { getTodos, USER_ID } from './api/todos';
-import classNames from 'classnames';
 import { Todo } from './types/Todo';
-
-type StatusType = 'all' | 'completed' | 'active';
+import { Header } from './components/Header';
+import { Footer } from './components/Footer';
+import { ErrorMessages } from './enums/ErrorMessages';
+import { StatusTypes } from './enums/StatusTypes';
+import { ErrorNotification } from './components/ErrorNotification';
+import { TodoList } from './components/TodoList';
 
 export const App: React.FC = () => {
-  const [errorMessage, setErrorMessage] = useState('');
+  const [errorMessage, setErrorMessage] = useState<ErrorMessages>(
+    ErrorMessages.NONE,
+  );
   const [loading, setLoading] = useState(false);
   const [todos, setTodos] = useState<Todo[]>([]);
-  const [statusFilter, setStatusFilter] = useState<StatusType>('all');
+  const [statusFilter, setStatusFilter] = useState<StatusTypes>(
+    StatusTypes.ALL,
+  );
 
   const todosCount = todos.reduce((count, todo) => count + +!todo.completed, 0);
 
@@ -22,7 +29,7 @@ export const App: React.FC = () => {
     getTodos()
       .then(setTodos)
       .catch(() => {
-        setErrorMessage('Unable to load todos');
+        setErrorMessage(ErrorMessages.GET_ERROR);
       })
       .finally(() => {
         setLoading(false);
@@ -32,7 +39,7 @@ export const App: React.FC = () => {
   useEffect(() => {
     if (errorMessage) {
       const timeoutId = setTimeout(() => {
-        setErrorMessage('');
+        setErrorMessage(ErrorMessages.NONE);
       }, 3000);
 
       return () => {
@@ -41,174 +48,38 @@ export const App: React.FC = () => {
     }
   }, [errorMessage]);
 
-  const visibleTodos = useMemo(() => {
-    return todos.filter(t =>
-      statusFilter === 'all'
-        ? true
-        : statusFilter === 'completed'
-          ? t.completed
-          : !t.completed,
-    );
-  }, [todos, statusFilter]);
   if (!USER_ID) {
     return <UserWarning />;
   }
-
 
   return (
     <div className="todoapp">
       <h1 className="todoapp__title">todos</h1>
 
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          {/* this button should have `active` class only if all todos are completed */}
-          <button
-            type="button"
-            className="todoapp__toggle-all active"
-            data-cy="ToggleAllButton"
-          />
-
-          {/* Add a todo on form submit */}
-          <form>
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-            />
-          </form>
-        </header>
+        <Header />
 
         {todos.length > 0 && (
           <>
-            <section className="todoapp__main" data-cy="TodoList">
-              {visibleTodos.map(todo => (
-                <div
-                  data-cy="Todo"
-                  className={classNames('todo', { completed: todo.completed })}
-                  key={todo.id}
-                >
-                  <label className="todo__status-label">
-                    <input
-                      data-cy="TodoStatus"
-                      type="checkbox"
-                      className="todo__status"
-                      checked={todo.completed}
-                    />
-                  </label>
+            <TodoList
+              todos={todos}
+              statusFilter={statusFilter}
+              loading={loading}
+            />
 
-                  <span data-cy="TodoTitle" className="todo__title">
-                    {todo.title}
-                  </span>
-
-                  <button
-                    type="button"
-                    className="todo__remove"
-                    data-cy="TodoDelete"
-                  >
-                    ×
-                  </button>
-
-                  {/* This form is shown instead of the title and remove button
-            <form>
-              <input
-                data-cy="TodoTitleField"
-                type="text"
-                className="todo__title-field"
-                placeholder="Empty todo will be deleted"
-                value="Todo is being edited now"
-              />
-            </form>
-            */}
-
-                  <div
-                    data-cy="TodoLoader"
-                    className={classNames('modal', 'overlay', {
-                      'is-active': loading,
-                    })}
-                  >
-                    <div
-                      className="modal-background
-                    has-background-white-ter"
-                    />
-                    <div className="loader" />
-                  </div>
-                </div>
-              ))}
-            </section>
-
-            <footer className="todoapp__footer" data-cy="Footer">
-              <span className="todo-count" data-cy="TodosCounter">
-                {todosCount} items left
-              </span>
-
-              {/* Active link should have the 'selected' class */}
-              <nav className="filter" data-cy="Filter">
-                <a
-                  href="#/"
-                  className={classNames('filter__link', {
-                    selected: statusFilter === 'all',
-                  })}
-                  data-cy="FilterLinkAll"
-                  onClick={() => setStatusFilter('all')}
-                >
-                  All
-                </a>
-
-                <a
-                  href="#/active"
-                  className={classNames('filter__link', {
-                    selected: statusFilter === 'active',
-                  })}
-                  data-cy="FilterLinkActive"
-                  onClick={() => setStatusFilter('active')}
-                >
-                  Active
-                </a>
-
-                <a
-                  href="#/completed"
-                  className={classNames('filter__link', {
-                    selected: statusFilter === 'completed',
-                  })}
-                  data-cy="FilterLinkCompleted"
-                  onClick={() => setStatusFilter('completed')}
-                >
-                  Completed
-                </a>
-              </nav>
-
-              {/* this button should be disabled if there are no completed todos */}
-              <button
-                type="button"
-                className="todoapp__clear-completed"
-                data-cy="ClearCompletedButton"
-              >
-                Clear completed
-              </button>
-            </footer>
+            <Footer
+              todosCount={todosCount}
+              statusFilter={statusFilter}
+              onStatusFilter={setStatusFilter}
+            />
           </>
         )}
       </div>
 
-      <div
-        data-cy="ErrorNotification"
-        className={classNames(
-          'notification',
-          'is-danger ',
-          'is-light',
-          'has-text-weight-normal',
-          { hidden: !errorMessage },
-        )}
-      >
-        <button
-          data-cy="HideErrorButton"
-          type="button"
-          className="delete"
-          onClick={() => setErrorMessage('')}
-        />
-        {errorMessage}
-      </div>
+      <ErrorNotification
+        errorMessage={errorMessage}
+        onErrorMessage={setErrorMessage}
+      />
     </div>
   );
 };
