+++ src/api/todos.ts
@@ -6,15 +6,3 @@ export const USER_ID = 3773;
 export const getTodos = (): Promise<Todo[]> => {
   return client.get<Todo[]>(`/todos?userId=${USER_ID}`);
 };
-
-// Add more methods here
-
-// Unable to load todos
-// <br />
-// Title should not be empty
-// <br />
-// Unable to add a todo
-// <br />
-// Unable to delete a todo
-// <br />
-// Unable to update a todo
