+++ src/components/ErrorNotification/ErrorNotification.scss
@@ -1,0 +1,12 @@
+
+.notification {
+  transition-property: opacity, min-height;
+  transition-duration: 1s;
+  min-height: 36px;
+}
+
+.notification.hidden {
+  min-height: 0;
+  opacity: 0;
+  pointer-events: none;
+}
