+++ src/components/Footer/Footer.scss
@@ -1,0 +1,47 @@
+.todoapp {
+  &__footer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    box-sizing: content-box;
+    height: 20px;
+    padding: 10px 15px;
+
+    font-size: 14px;
+
+    color: #777;
+    text-align: center;
+    border-top: 1px solid #e6e6e6;
+
+    box-shadow: 0 1px 1px rgba(0, 0, 0, 0.2),
+      0 8px 0 -3px #f6f6f6,
+      0 9px 1px -3px rgba(0, 0, 0, 0.2),
+      0 16px 0 -6px #f6f6f6,
+      0 17px 2px -6px rgba(0, 0, 0, 0.2);
+  }
+}
+
+.filter {
+  display: flex;
+
+  &__link {
+    margin: 3px;
+    padding: 3px 7px;
+
+    color: inherit;
+    text-decoration: none;
+
+    border: 1px solid transparent;
+    border-radius: 3px;
+
+    &:hover {
+      border-color: rgba(175, 47, 47, 0.1);
+    }
+
+    &.selected {
+      border-color: rgba(175, 47, 47, 0.2);
+    }
+  }
+}
+
