+++ src/components/Header/Header.scss
@@ -1,0 +1,5 @@
+ .todoapp {
+   &__header {
+     position: relative;
+   }
+ }
