+++ src/components/Header/Header.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import './Header.scss';
+
+export const Header = () => {
+  return (
+    <header className="todoapp__header">
+      {/* this button should have `active` class only if all todos are completed */}
+      <button
+        type="button"
+        className="todoapp__toggle-all active"
+        data-cy="ToggleAllButton"
+      />
+
+      {/* Add a todo on form submit */}
+      <form>
+        <input
+          data-cy="NewTodoField"
+          type="text"
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+        />
+      </form>
+    </header>
+  );
+};
