+++ src/components/TodoList/TodoList.scss
@@ -1,0 +1,5 @@
+.todoapp {
+  &__main {
+    border-top: 1px solid #e6e6e6;
+  }
+}
