+++ src/enums/ErrorMessages.tsx
@@ -1,0 +1,8 @@
+export enum ErrorMessages {
+  NONE = '',
+  EMPTY_TITLE = 'Title should not be empty',
+  GET_ERROR = 'Unable to load todos',
+  CREATE_ERROR = 'Unable to add a todo',
+  DELETE_ERROR = 'Unable to delete a todo',
+  UPDATE_ERROR = 'Unable to update a todo',
+}
