+++ src/enums/StatusTypes.tsx
@@ -1,0 +1,5 @@
+export enum StatusTypes {
+  ALL = 'all',
+  COMPLETED = 'completed',
+  ACTIVE = 'active',
+}
