--- src/styles/filter.scss
@@ -1,22 +1,0 @@
-.filter {
-  display: flex;
-
-  &__link {
-    margin: 3px;
-    padding: 3px 7px;
-
-    color: inherit;
-    text-decoration: none;
-
-    border: 1px solid transparent;
-    border-radius: 3px;
-
-    &:hover {
-      border-color: rgba(175, 47, 47, 0.1);
-    }
-
-    &.selected {
-      border-color: rgba(175, 47, 47, 0.2);
-    }
-  }
-}
