+++ src/styles/index.scss
@@ -8,18 +8,4 @@ body {
   margin: 0 auto;
 }
 
-.notification {
-  transition-property: opacity, min-height;
-  transition-duration: 1s;
-  min-height: 36px;
-}
-
-.notification.hidden {
-  min-height: 0;
-  opacity: 0;
-  pointer-events: none;
-}
-
 @import "./todoapp";
-@import "./todo";
-@import "./filter";
