+++ src/styles/todoapp.scss
@@ -23,10 +23,6 @@
     text-rendering: optimizeLegibility;
   }
 
-  &__header {
-    position: relative;
-  }
-
   &__toggle-all {
     position: absolute;
 
@@ -76,32 +72,6 @@
       font-weight: 300;
       color: #e6e6e6;
     }
-  }
-
-  &__main {
-    border-top: 1px solid #e6e6e6;
-  }
-
-  &__footer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    box-sizing: content-box;
-    height: 20px;
-    padding: 10px 15px;
-
-    font-size: 14px;
-
-    color: #777;
-    text-align: center;
-    border-top: 1px solid #e6e6e6;
-
-    box-shadow: 0 1px 1px rgba(0, 0, 0, 0.2),
-      0 8px 0 -3px #f6f6f6,
-      0 9px 1px -3px rgba(0, 0, 0, 0.2),
-      0 16px 0 -6px #f6f6f6,
-      0 17px 2px -6px rgba(0, 0, 0, 0.2);
   }
 
   &__clear-completed {
